Update zm_create.sql
On a fresh ZoneMinder install, there's a default zone preset active which works pretty well. 

If the user goes to the zone settings and plays around with the selectable presets, he looses the default setting forever, if didn't take notes by hand on a piece of paper.

This commit simply adds to the top of the presets list the default value which comes with the original fresh install.