Update zmupdate.pl.in to give option to convert to InnoDB tables
When upgrading the zoneminder database to 1.26, this proposed change will ask the user to convert existing MyISAM tables over InnoDB.  It first scans for existing MyISAM tables and will skip over any tables that are already converted to InnoDB.

I've tested this on (a copy of) a production database server and have verified it doesn't alter existing tables from a different schema.

And I'm done messing with the zmupdate script. I promise.
I understand if this change is considered outside the scope of a current bug fix release.