Update getBrowser() to match IE10
Updates the preg_match expression to match Internet Explorer 10. The previous pattern would not detect IE10, and hence zoneminder would not attempt to stream via Cambozola, which leaves IE10 users looking at a broken link.

NOTE: There is another file in the zoneminder source tree, AssetDispatcher.php, that also uses the same pattern match expression.  It is part of the Cake folder, which I am not familiar with.  Please advise if we should edit that file as well.