Fix "Can't stat : No such file or directory" message
Many zoneminder systems have a "Can't stat : No such file or directory" in their logs each time zmfix is run. The problem was traced back to the sql query, which looked for a ControlDevice field that was not null.  However, the ControlDevice field is empty, rather than null, when there is no control device.