Update zm_create.sql.in
Don't explicitly set the Id number of each ptz control.  The auto increment flag has previously been turned on so mysql will insert the proper value for us when we create a new record. This will make the list of ptz controls easier to manager as the list gets bigger.