Update zmtrigger.pl.in
zmtrigger.pl should not open serial port #1 by default. There could be situations where other hardware etc. would be attached to the port that is not used by zoneminder. 

Better if this can be commented out by default and if a user needs serial access, he can enable it by removing the comment.

This line also gives an permissions error when running on Ubuntu which prevents zmtrigger to start. That error can probably be fixed by adding the zm user permission to the serial port. But the main issue is that if the program does not need access to the port, it should not open it.