Login screen instead of error
If a view returns an error and the user is not logged in, the user gets
a login prompt which auto redirects him to the originally intended
location after he logs in.  Postlogin.js.php now redirects to the
ZoneMinder URL with any specified GET query string appended to the URL.
 The query string is saved as a hidden field in the login screen so
that after the user logs in, he gets redirected to his original
destination instead of the ZM console.

If the user is logged in and there is an error, then the user gets the
original error screen.

This change allows a person to click a ZoneMinder event e-mail link,
enter his username and password, and then see the intended event
without having to access the ZoneMinder front page separately to log in
first.

Changed a couple stray tabs to spaces in login.php.