Offer login instead of error
When a user accesses a view but receives an error, and is not logged
on, he is offered a login prompt instead.  The login prompt saves the
original query URL in a hidden field, and postlogin redirects back to
the original URL once the user has logged on.

If the user is logged in and there is an error, no login prompt is
shown.

This allows the user to click an event link in an e-mail and then log
in before being shown the event, instead of requiring going back
through the front ZM page to log in.