Fix bug:  'mid' was not properly set
Since migrating to ui-router, the monitor's ID was not properly set
when trying to edit an existing monitor.  This caused existing
monitors to not load properly when in the edit / detail page.

Now, the monitor id (mid) is set as a state parameter, and is
accessed from $state.params.mid