Add zone view / edit functionality to new ui
You should now be able to see a monitor's zones by clicking on the
respective 'Edit' link.  From there, you'll see a list view of that
monitor's Zones.

From there, you can click on the Zone name, which will bring up the
details for that zone.

You can not yet save zones, and not all of the values for that zone
are displayed.