Add CakePHP Crud plugin to the API
I wish I knew about this plugin months ago.

By default PDO and CakePHP cast int as string.  This prevents me
from using `<input type="number">`, as the modle that the input
is bound to is actually a string, not an int.

Crud by default casts int back to int via ApiTransformationListener

It also has a bunch of other features which makes the life of a
REST / CRUD developer easier.