Adding continuous movement methods and fixing reset
This patch adds the continuous movement methods provided in the patch
for Issue #2414. Note that these are not truly continuous as they
result in movement for less than a ms. Adapting the code to support
truly continuous movement would invovle some considerable work.

This patch also correct the reset method and adds a reboot method.
The two are different creatures with different effects.

POD added as well.