Removing redundant get request code
Two slightly different versions of code are currently used to post
the get request to the camera in order to execute commands. This
patch modifies the open method in order to allow its re-use and
removes redundant code.

Note: This is the first installment on changes towards improving
the way the HTTP transactions are handled.