Delete the selected event, not the most recent one (#2579)
The click event uses a closure to call deleteEvent with event.Id as parameter. At the end of the loop the value of event.Id is always equal to the id of the most recent event, so trying to delete any row always results in deleting the most recent event. 

Making the event block scoped instead of function scoped fixes this.