Fix iOS autocapitalizing username field on login
I propose removing the auto-capitalization from the username field for Safari under iOS by adding autocapitalize="none" to the username form field. Usernames rarely start with a capital letter, so I think this would be a usability improvement for users logging in through iPhones or iPads.

Having to login to ZM under iOS, I've been frustrated at the need to always press on the Caps Lock key on the virtual keyboard to disable the capitalization of the first letter. This is because iOS auto capitalized non-password and non-email HTML form fields. ZM is also case sensitive, so "Admin" will not work if the main user is "admin".