Allows users to set an specific "Onvif alarm text" for each monitor
Updated es_la.php language file.
With this personalization the cameras that use Alarm strings different to "MotionAlarm"
such as Dahua, will be able to process Onvif alarms. To make it easier, the change
sets a default "Alarm Text"  value for it to work out of the box and allows to modify it
acording to the particular necesities of each brand. It consists of a new column on
table Monitors, changes on Monitors.h, Monitors.cpp and a change on UI to manage
the value of Alarm Text per camera.