fix(event): Honor the Alt key with left/right arrow
Fixes a race condition when using Alt-LeftArrow to go back to the previous page (standard browser behavior).

Sometimes it would go the previous event and sometimes it would go to the previous page.  Now it will consistently go to the previous page when the alt key is pressed with the left arrow.