Event table data should be aligned with column headers
This is more noticeable when some columns of the events table are toggled off, but the column headers were centered and the data in columns were left justified.  This made it difficult to read which column went to which data.

This change centers both the column headers and the data in the columns so that they line up.