Refresh table on show
After opening an event from the events table and editing some properties, returning back to the events table would show the cached version of the table without the new values.  

This change will refresh the events table each time it's shown. 

It is a bit of a performance hit, but it's more important to show accurate information after a quick delay than to show bad/inaccurate information fast.