Fis style in Options page
 * Why extra padding at top? 
  padding-top: 2rem;
  */

I don't know the answer, because... I didn't add it :)
Before my edit, it was generally “4rem”; it seemed like a lot to me, so I halved it.
I didn’t clean it up completely because I was careful not to break something completely.

I understood why. If the <form> does not have buttons at the top, then there will be no padding between the form and the "id="optionsContainer" (for example the "Storage" section).
For now, so be it.