Add wrapper fo scroll table & remove padding-top from "form" & add padding to #optionsContainer
When sticky mode is enabled, the table must be scrolled vertically if it does not fit on the page.
It may not be the best option, but now it is necessary. For example, for the pages "Servers, Storage, Groups", etc.

Changing the padding-top is necessary to maintain a single style if there are some other elements in front of the form (buttons for example) or there is nothing in front of the form.