Fix: Panning didn't work when using "preventDefault" (panzoom.js)
This happened after PR #4137
When Scale panZoom > 1, you need to use "preventDefault" for Panning to work correctly. But at the same time (if Scale panZoom > 1), clicking on the progress bar generated by video.js will not work.
In principle, this is not critical, because when increasing, the progress bar generated by video.js usually goes beyond the visible boundaries.