Don't show video player control panel unless mouse cursor is over player (skin.js)
Relevant for the situation when during real-time viewing using RTSP2Web type MSE the stream lag time is adjusted. 
For example, on the Montage page the periodic appearance of the control panel is very annoying. 
Allow the panel to appear only when the cursor is hovered over the video object.