Control RTCPeerConnection and video codec support, stop all monitors when closing the page, etc. (MonitorStream.js)
Added this.name and this.capturing to monitor stream
this.capturing is not used yet.

If the browser does not support "RTCPeerConnection", then warn in the console, avoid errors and repeated restarts

When closing the page, not only close the Web socket for MSE, but also stop all monitors. This is more correct, since it will allow you to avoid a number of errors when closing the page.

If the codec is not supported by the browser - there is no point in performing repeated restarts. Just stop.