Added "preload" for "material-icons.woff2", added parameters to the call "output_link_if_exists" (functions.php)
Now material-icons.woff2 starts loading without the slightest delay. 99.99% of the time is spent loading the file.
material-icons.woff is not preloaded, because if it is loaded simultaneously with material-icons.woff2, then a lot of time is wasted.
If you use JS to determine the browser's support for woff2, and then preload, it does not work. The font will be loaded in normal mode, without preload.
Almost all modern browsers support woff2.
This option is the fastest.